requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
PyMySQL==1.1.0
Werkzeug==2.3.7
cryptography>=3.4.8
python-dotenv>=0.19.0